+++ portal/tests/snapshots/snap_test_partials.py
@@ -7,7 +7,9 @@ from snapshottest import Snapshot
 
 snapshots = Snapshot()
 
-snapshots['test_banner 1'] = '''<div class="banner banner--teacher">
+snapshots[
+    "test_banner 1"
+] = """<div class="banner banner--teacher">
     <div class="container">
         <div class="row">
             <div class="col-sm-12 d-flex">
@@ -18,6 +20,7 @@ snapshots['test_banner 1'] = '''<div class="banner banner--teacher">
                     
                     
                         <p>Test text</p>
+                    
                     
                 </div>
                 <div>
@@ -33,9 +36,11 @@ snapshots['test_banner 1'] = '''<div class="banner banner--teacher">
         </div>
     </div>
 </div>
-'''
+"""
 
-snapshots['test_benefits 1'] = '''
+snapshots[
+    "test_benefits 1"
+] = """
 
 <div class="grid-benefits col-sm-8 col-center">
     
@@ -75,9 +80,11 @@ snapshots['test_benefits 1'] = '''
         </div>
     
 </div>
-'''
+"""
 
-snapshots['test_card_list 1'] = '''
+snapshots[
+    "test_card_list 1"
+] = """
 
 
 <div class="grid grid-worksheets grid__fit col-sm-8 col-center">
@@ -175,9 +182,11 @@ snapshots['test_card_list 1'] = '''
         </div>
     
 </div>
-'''
+"""
 
-snapshots['test_character_list 1'] = '''
+snapshots[
+    "test_character_list 1"
+] = """
 
 
 <div class="grid grid-characters grid__fit">
@@ -201,26 +210,19 @@ snapshots['test_character_list 1'] = '''
         </div>
     
 </div>
-'''
+"""
 
-snapshots['test_game_banner 1'] = '''
+snapshots[
+    "test_headline 1"
+] = """<section>
+    <h4>Test title</h4>
+</section>
+<p class="container">Test description</p>
+"""
 
-<div class="banner--game col-center col-lg-8 col-sm-12 test--class">
-    <div>
-        <h4>Test ages</h4>
-        <h2>Test title</h2>
-        <p class="banner--game__text"><strong>Test description</strong></p>
-        <a href="/play/"
-           class="button button--home button-right-arrow">Test button</a>
-    </div>
-</div>
-'''
-
-snapshots['test_headline 1'] = '''<h4>Test title</h4>
-<div class="col-sm-6 col-center">Test description</div>
-'''
-
-snapshots['test_hero_card 1'] = '''
+snapshots[
+    "test_hero_card 1"
+] = """
 
 
 <div class="card col-sm-8 col-center">
@@ -229,8 +231,8 @@ snapshots['test_hero_card 1'] = '''
         <h4 class="card__title">Test title</h4>
         <p>Test description</p>
         <div class="button-group button-group__icon">
-            <a href="/materials/test_pdf_name" class="button button--primary button-right-arrow">
-                Test button 1
+            <a target="_blank" href="https://www.codeforlife.education" class="button button--primary button--icon">
+                Test button 1<span class="iconify" data-icon="mdi:open-in-new"></span>
             </a>
             <a href="/kurono/play/1/" class="button button--primary button-right-arrow">
                 Test button 2
@@ -238,4 +240,4 @@ snapshots['test_hero_card 1'] = '''
         </div>
     </div>
 </div>
-'''
+"""