+++ portal/views/teacher/teach.py
@@ -2,7 +2,8 @@ from __future__ import division
 
 import csv
 import json
-from datetime import timedelta
+from datetime import timedelta, datetime
+from enum import Enum
 from functools import partial, wraps
 from uuid import uuid4
 
@@ -15,7 +16,7 @@ from common.helpers.generators import (
     generate_password,
     get_hashed_login_id,
 )
-from common.models import Class, Student, Teacher
+from common.models import Class, Student, Teacher, DailyActivity
 from common.permissions import logged_in_as_teacher
 from django.contrib import messages as messages
 from django.contrib.auth.decorators import login_required, user_passes_test
@@ -53,19 +54,6 @@ REMINDER_CARDS_PDF_COLUMNS = 1
 REMINDER_CARDS_PDF_WARNING_TEXT = (
     "Please ensure students keep login details in a secure place"
 )
-
-
-@login_required(login_url=reverse_lazy("teacher_login"))
-@user_passes_test(logged_in_as_teacher, login_url=reverse_lazy("teacher_login"))
-def default_solution(request, levelName):
-    if 80 <= int(levelName) <= 91:
-        return render(
-            request, "portal/teach/teacher_solutionPY.html", {"levelName": levelName}
-        )
-    else:
-        return render(
-            request, "portal/teach/teacher_solution.html", {"levelName": levelName}
-        )
 
 
 @login_required(login_url=reverse_lazy("teacher_login"))
@@ -797,6 +785,11 @@ def process_move_students_form(request, formset, old_class, new_class):
     )
 
 
+class DownloadType(Enum):
+    CSV = 1
+    LOGIN_CARDS = 2
+
+
 @login_required(login_url=reverse_lazy("teacher_login"))
 @user_passes_test(logged_in_as_teacher, login_url=reverse_lazy("teacher_login"))
 def teacher_print_reminder_cards(request, access_code):
@@ -906,6 +899,9 @@ def teacher_print_reminder_cards(request, access_code):
     compute_show_page_end(p, x, y)
 
     p.save()
+
+    count_student_details_click(DownloadType.LOGIN_CARDS)
+
     return response
 
 
@@ -920,13 +916,21 @@ def teacher_download_csv(request, access_code):
     if klass.teacher.new_user != request.user:
         raise Http404
 
+    class_url = request.build_absolute_uri(
+        reverse("student_login", kwargs={"access_code": access_code})
+    )
+
     # Use data from the query string if given
     student_data = get_student_data(request)
     if student_data:
         writer = csv.writer(response)
-        writer.writerow([access_code])
+        writer.writerow([access_code, class_url])
         for student in student_data:
-            writer.writerow([student["name"], student["login_url"]])
+            writer.writerow(
+                [student["name"], student["password"], student["login_url"]]
+            )
+
+    count_student_details_click(DownloadType.CSV)
 
     return response
 
@@ -949,6 +953,19 @@ def compute_show_page_character(p, x, y, NUM_Y):
 def compute_show_page_end(p, x, y):
     if x != 0 or y != 0:
         p.showPage()
+
+
+def count_student_details_click(download_type):
+    activity_today = DailyActivity.objects.get_or_create(date=datetime.now().date())[0]
+
+    if download_type == DownloadType.CSV:
+        activity_today.csv_click_count += 1
+    elif download_type == DownloadType.LOGIN_CARDS:
+        activity_today.login_cards_click_count += 1
+    else:
+        raise Exception("Unknown download type")
+
+    activity_today.save()
 
 
 def invite_teacher(request):